ADDED FUNCTIONALITY TO CART PAGE
Added functionality to the cart page so when we are clicking the Remove button of an individual cart item the count of the quantity is getting decreased by 1. If the quantity is already 0 when we are clicking the remove button an alert is shown "Cart item is already Empty cannot delete item".