ADDED THE CONNECTIVITY BETWEEN FRONTEND AND BACKEND
Here we have added the functionality where we are able to add users to the database and also check the user from the database from the frontend and also set up proper http status code to be send in case of any error so that the front end can catch the error and show in the logs.
IMP : for now I have hardcoded the frontend to backend Rest API calls.